edit/테스트용 추천 시스템 csv파일 및 서버에 작성한 코드 파일 삭제, 서비스함수 recommendation앱에 분리
