+++ recommendation/services/recomendation_service.py
@@ -1,0 +1,27 @@
+from django.http import HttpResponse
+
+from recommendation import recommender
+from main_page.models import Letter as LetterModel
+
+
+
+
+def recommend_worryboard_list(cur_user : object):
+    """
+    추천시스템 service
+    """
+    try:
+        latest_user_letter = LetterModel.objects.filter(letter_author=cur_user).order_by(
+            "-create_date"
+        )[:1]
+        worryboard_id_of_letter = latest_user_letter[0].worryboard.id
+        recomendation_sys = recommender.recommend_worryboard
+        final_worryboard_list = recomendation_sys.recommend_worries(
+            worryboard_id_of_letter, cur_user
+        )
+        return final_worryboard_list
+    
+    except IndexError:
+        print("네가 쓴 편지가 없어서 키에러!!")
+        return HttpResponse(status=204)
+
