--- test_cosine_sim.csv
@@ -1,8 +1,0 @@
-0,1,2,3,4,5,6
-0.0,0.0,0.0,0.0,0.0,0.0,0.0
-0.0,1.0000000000000002,0.0,0.0,0.6387085483562188,0.0,0.0
-0.0,0.0,1.0,1.0,0.0,0.0,0.0
-0.0,0.0,1.0,1.0,0.0,0.0,0.0
-0.0,0.6387085483562188,0.0,0.0,1.0,0.0,0.0
-0.0,0.0,0.0,0.0,0.0,1.0,1.0
-0.0,0.0,0.0,0.0,0.0,1.0,1.0
