--- test_worryboard.csv
@@ -1,7 +1,0 @@
-id,content,author_id,category_id,create_date
-1,너무 더운데 뭘 먹을까요,1,1,2022-07-30 10:18:04.374540
-2,사랑이 뭔가요,2,2,2022-07-30 10:20:28.524279
-3,연애를 하고 싶어요,2,2,2022-07-30 10:20:40.499266
-4,사랑할 수 있을까요,3,2,2022-07-30 10:21:33.004707
-5,아이가 밤마다 울어요,3,6,2022-07-30 10:21:50.288744
-
